refactor: mcp store split from settings store
