+++ src/renderer/src/stores/settings.ts
@@ -169,14 +169,6 @@ export const useSettingsStore = defineStore('settings', () => {
     }
   }
 
-  // MCP 安装缓存
-  const mcpInstallCache = ref<string | null>(null)
-
-  // 清理 MCP 安装缓存
-  const clearMcpInstallCache = () => {
-    mcpInstallCache.value = null
-  }
-
   // 监听 deeplink 事件
   window.electron.ipcRenderer.on(DEEPLINK_EVENTS.MCP_INSTALL, async (_, data) => {
     const { mcpConfig } = data
@@ -211,7 +203,7 @@ export const useSettingsStore = defineStore('settings', () => {
 
     // 存储 MCP 配置数据到缓存
     if (data) {
-      mcpInstallCache.value = mcpConfig
+      mcpStore.setMcpInstallCache(mcpConfig)
     }
   })
 
@@ -1253,7 +1245,6 @@ export const useSettingsStore = defineStore('settings', () => {
     await configP.setLoggingEnabled(enabled)
   }
 
-
   ///////////////////////////////////////////////////////////////////////////////////////
   const setCopyWithCotEnabled = async (enabled: boolean) => {
     copyWithCotEnabled.value = Boolean(enabled)
@@ -1451,8 +1442,6 @@ export const useSettingsStore = defineStore('settings', () => {
     testSearchEngine,
     refreshSearchEngines,
     findModelByIdOrName,
-    mcpInstallCache,
-    clearMcpInstallCache,
     isUpdating: upgradeStore.isUpdating,
     loadSavedOrder,
     updateProvidersOrder,
